mm-video: vdec: Update output buffer count in thumbnail mode
This change updates the output buffer requirements when thumbnail mode
is enabled. In thumbnail mode the required output buffers equal to 1.
Output Buffer required for thumbnail mode were not updated earlier from
the driver as a result of which allocating memory for default buffer
count failed and thumbnail were not shown for some file formats.

Change-Id: I607a5b0e3bfcad8056ff71557db50d0e268cf348
CRs-Fixed: 513180